+++ bluer/examples/discover_devices.rs
@@ -1,6 +1,6 @@
 //! Discover Bluetooth devices and list them.
 
-use bluer::{Adapter, AdapterEvent, Address, DeviceEvent};
+use bluer::{Adapter, AdapterEvent, Address, DeviceEvent, DiscoveryFilter, DiscoveryTransport};
 use futures::{pin_mut, stream::SelectAll, StreamExt};
 use std::{collections::HashSet, env};
 
@@ -35,13 +35,28 @@ async fn query_all_device_properties(adapter: &Adapter, addr: Address) -> bluer:
 async fn main() -> bluer::Result<()> {
     let with_changes = env::args().any(|arg| arg == "--changes");
     let all_properties = env::args().any(|arg| arg == "--all-properties");
+    let le_only = env::args().any(|arg| arg == "--le");
+    let br_edr_only = env::args().any(|arg| arg == "--bredr");
     let filter_addr: HashSet<_> = env::args().filter_map(|arg| arg.parse::<Address>().ok()).collect();
 
     env_logger::init();
     let session = bluer::Session::new().await?;
     let adapter = session.default_adapter().await?;
-    println!("Discovering devices using Bluetooth adapater {}\n", adapter.name());
+    println!("Discovering devices using Bluetooth adapter {}\n", adapter.name());
     adapter.set_powered(true).await?;
+
+    let filter = DiscoveryFilter {
+        transport: if le_only {
+            DiscoveryTransport::Le
+        } else if br_edr_only {
+            DiscoveryTransport::BrEdr
+        } else {
+            DiscoveryTransport::Auto
+        },
+        ..Default::default()
+    };
+    adapter.set_discovery_filter(filter).await?;
+    println!("Using discovery filter:\n{:#?}\n\n", adapter.discovery_filter().await);
 
     let device_events = adapter.discover_devices().await?;
     pin_mut!(device_events);
@@ -57,7 +72,7 @@ async fn main() -> bluer::Result<()> {
                             continue;
                         }
 
-                        println!("Device added: {}", addr);
+                        println!("Device added: {addr}");
                         let res = if all_properties {
                             query_all_device_properties(&adapter, addr).await
                         } else {
@@ -74,15 +89,15 @@ async fn main() -> bluer::Result<()> {
                         }
                     }
                     AdapterEvent::DeviceRemoved(addr) => {
-                        println!("Device removed: {}", addr);
+                        println!("Device removed: {addr}");
                     }
                     _ => (),
                 }
                 println!();
             }
             Some((addr, DeviceEvent::PropertyChanged(property))) = all_change_events.next() => {
-                println!("Device changed: {}", addr);
-                println!("    {:?}", property);
+                println!("Device changed: {addr}");
+                println!("    {property:?}");
             }
             else => break
         }